Add notion of backend, implement MemoryBackend
This is a functionality-neutral change that adds the notion of a
"backend" that acts as an observer to the main data and connection
path. This is marching towards persistence, as the backend is expected
to accept non-blocking notifications for new/deleted VMs. The backend
is now responsible for tracking changes, handling the admin
connections, and firing off change "hooks". In the MemoryBackend, the
hooks do nothing but log.