Add backend logging hooks
A big part of our use case for vSPC.py is logging the activity on the
consoles of our virtualized Linux machines. We implement logging with a
custom backend, which helps keep the network processing path small and
simple. This requires a new hook, called when messages are received from
remote VMs.