Handle select.error in run_once.
Our logging backend listens for SIGHUP so it knows when to reload its
log files. This potentially changes the disposition of SIGHUP (from
ignore to something else), which potentially breaks the select.select in
run_once (by interrupting a syscall). This results in a select.error. To
address this, we catch select.error and return without doing anything.
run_forever is currently the only caller of run_once, and will
unconditionally call run_once again after it returns, so this change
doesn't alter the way data are processed in the current program. To
allow future callers of run_once to distinguish between run_once
completing without a select.error and our new path, we return False if
we return because of a select.error.