Change query socket handler to take a vspc object as an argument.
In the future, we may want to register a client socket as a client
within the vSPC, so it is read from and written to without intervention
by the backend. We'll need a reference to the vSPC for this to work. An
alternative would be to seed the backend with the vSPC reference upon
creation, but I'm not sure I like the implications of that; you'd have
to have the vSPC available when creating the backend, and you could only
use the backend with the one vSPC.