Convert vSPC to use a worker thread to process connections.
This is necessary to allow the admin backend to interact with the rest
of the framework in the way that we want. Reader events are now
processed using a single worker thread and a thread-safe Queue.Queue
instance. The Selector calls methods that write tasks to this queue, and
the worker thread pops the tasks off one at a time and performs them.
Changes to the state of the vSPC are all performed by the worker thread,
so we don't need to redesign the vSPC class to be thread-safe. There
aren't any major functional changes with these changes; the only
noticeable difference is that there can be a slight delay between a
connection to the local listener port and when the selector starts
listening for new data on that connection.