Tweak admin protocol to return strings of seed data instead of lists.
This allows us to more accurately reason about scrollback limits.
Before, we limited the number of messages. A message can correspond to a
single character or a line or two, depending on how long it takes the
vSPC to read a vm connection and other factors. This makes the message
limit introduced earlier a bit imprecise; we might be limiting ourselves
to 25 characters or 25 lines, depending on these factors. This commit
changes scrollback to be a string, which makes the configuration knob
more precise.