Remove fd from the epoll stuff before closing
After we started closing old sockets, we were unable to add new sockets to the
epoller.  This is because we weren't removing anything from the epoller or our
poller structures, so if a repeat fileno came in we didn't do work to add it
but were surprised when it wasn't set up correctly.  To fix this, we actually
pull the socket out of the structures when we close it.