Clean up state management in Poller; move state to PollEventSource.
A PollEventSource object is responsible for managing all state related
to using epoll with a stream. In particular, it manages the current
epoll mask, stores the fileno and stream, and stores read and write
handlers. This removes the need to add multiple overlapping maps to
Poller when we want to track something else. No changes to public (i.e.,
used) API, no functional changes.